move access check to module
