+++ controller/user/user.controller.js
@@ -1,5 +1,6 @@
 var async = require("async");
-var dateFormat = require('dateformat');
+var dateFormat = require("dateformat");
+var checkAccess = require("../../js/checkAccess.js");
 
 var validator = function(req, res, next) {
 	
@@ -16,85 +17,6 @@ var validator = function(req, res, next) {
 	}
 	*/
 	next();
-}
-
-var accessLevels=["none","view","edit","add","remove"];
-
-var checkAccessToMRI = function checkAccessToMRI(mri, projects, user, access) {
-    var p, requestedLevel = accessLevels.indexOf(access);
-    
-    for(p in projects) {
-        if(projects[p].files.list.indexOf(mri.source)>=0) {
-            console.log(mri.source,user,access,projects[p].files.list.indexOf(mri.source));
-            var publicLevel = accessLevels.indexOf(projects[p].files.access.public);
-            var c, collaborators;
-            
-            // check if user has owner access
-            if(user === projects[p].files.access.owner) {
-                continue;
-            }
-            
-            // check if user has collaborator access
-            var accessOk = false;
-            collaborators=projects[p].files.access.whitelist;
-            for(c in collaborators) {
-                if(collaborators[c].userId === user) {
-                    var collaboratorAccessLevel = accessLevels.indexOf(collaborators[c].access);
-                    if(requestedLevel>collaboratorAccessLevel) {
-                        console.log("Collaborator access refused from project",projects[p].shortname);
-                        return false;
-                    } else {
-                        accessOk = true;
-                    }
-                    break;
-                }
-            }
-            if(accessOk) {
-                continue;
-            }
-
-            // check if user has public access
-            if(requestedLevel>publicLevel) {
-                console.log("Public access refused from project",projects[p].shortname);
-                return false;
-            }
-        }
-    }
-    
-    return true;
-}
-var checkAccessToProject = function checkAccessToProject(project, user, access) {
-    var p, requestedLevel = accessLevels.indexOf(access);
-    
-    var publicLevel = accessLevels.indexOf(project.files.access.public);
-    var c, collaborators;
-    
-    // check if user has owner access
-    if(user === project.files.access.owner) {
-        return true;
-    }
-    
-    // check if user has collaborator access
-    collaborators=project.files.access.whitelist;
-    for(c in collaborators) {
-        if(collaborators[c].userId === user) {
-            var collaboratorAccessLevel = accessLevels.indexOf(collaborators[c].access);
-            if(requestedLevel>collaboratorAccessLevel) {
-                console.log("Collaborator access refused to project",project.shortname);
-                return false;
-            } else {
-                return true;
-            }
-        }
-    }
-
-    // check if user has public access
-    if(requestedLevel>publicLevel) {
-        console.log("Public access refused to project",project.shortname);
-        return false;
-    }
-    
-    return true;
 }
 
 var user = function(req, res) {
@@ -126,13 +48,13 @@ var user = function(req, res) {
                     
                     // filter for view access
                     for(i in unfilteredMRI)
-                        if(checkAccessToMRI(unfilteredMRI[i],unfilteredProjects,loggedUser,"view"))
+                        if(checkAccess.toMRI(unfilteredMRI[i],unfilteredProjects,loggedUser,"view"))
                             mri.push(unfilteredMRI[i]);
                     for(i in unfilteredAtlas)
-                        if(checkAccessToMRI(unfilteredAtlas[i],unfilteredProjects,loggedUser,"view"))
+                        if(checkAccess.toMRI(unfilteredAtlas[i],unfilteredProjects,loggedUser,"view"))
                             atlas.push(unfilteredAtlas[i]);
                     for(i in unfilteredProjects)
-                        if(checkAccessToProject(unfilteredProjects[i],loggedUser,"view"))
+                        if(checkAccess.toProject(unfilteredProjects[i],loggedUser,"view"))
                             projects.push(unfilteredProjects[i]);
                     console.log(values[0].length-mri.length,"MRIs filtered");
                     console.log(values[1].length-atlas.length,"atlases filtered");
